chore(tools): inline the now-single-caller Reflect helper and report an unparsed parameter once
Review follow-ups. shared/reflect-method.ts was shared because two rules
read it; the merge left it with one caller, so it lives in no-reflect.ts.
The merged parameter rule classified an annotation at each of two report
sites; it now classifies once and reports once, which keeps the `cause`
exemption visibly scoped to the unknown branch.
